feat: Remove Hilt completely — kotlin-inject is sole DI system (Phase 3.3)
Pure deletion PR — no business logic changes:

- Remove @HiltAndroidApp from GarageApplication
- Remove @AndroidEntryPoint from MainActivity
- Delete all 15 Hilt @Module classes (dead code since kotlin-inject provides everything)
- Remove Hilt Gradle plugin and dependencies (hilt-android, hilt-navigation-compose, hilt-compiler)
- Remove all dagger/javax.inject imports
- Remove @Inject from UseCases (provided via @Provides in AppComponent)
- Rename files to match single-class convention (DoorRepository → DoorRepositoryImpl)

Zero dagger/hilt references remain in source code.
