+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/auth/AuthRepository.kt
@@ -30,10 +30,6 @@ import com.chriscartland.garage.domain.repository.AuthRepository
 import com.google.firebase.Firebase
 import com.google.firebase.auth.GoogleAuthProvider
 import com.google.firebase.auth.auth
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -41,135 +37,124 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.suspendCancellableCoroutine
-import javax.inject.Inject
 import kotlin.coroutines.resume
 import kotlin.coroutines.suspendCoroutine
 
 const val RC_ONE_TAP_SIGN_IN = 1
 
-class AuthRepositoryImpl
-    @Inject
-    constructor(
-        private val appLoggerRepository: AppLoggerRepository,
-    ) : AuthRepository {
-        private val _authState = MutableStateFlow<AuthState>(AuthState.Unknown)
-        override val authState: StateFlow<AuthState> = _authState.asStateFlow()
+class AuthRepositoryImpl(
+    private val appLoggerRepository: AppLoggerRepository,
+) : AuthRepository {
+    private val _authState = MutableStateFlow<AuthState>(AuthState.Unknown)
+    override val authState: StateFlow<AuthState> = _authState.asStateFlow()
 
-        init {
-            CoroutineScope(Dispatchers.IO).launch {
-                refreshFirebaseAuthState()
-            }
-        }
-
-        /**
-         * Sign in with Google ID Token and update the [authState].
-         */
-        override suspend fun signInWithGoogle(idToken: GoogleIdToken): AuthState {
-            firebaseSignInWithGoogle(idToken)
-            return refreshFirebaseAuthState()
-        }
-
-        /**
-         * Use the Google ID Token to sign in with Firebase.
-         */
-        private suspend fun firebaseSignInWithGoogle(idToken: GoogleIdToken) {
-            suspendCoroutine { continuation ->
-                Log.d(TAG, "firebaseAuthWithGoogle")
-                val credential = GoogleAuthProvider.getCredential(idToken.asString(), null)
-                Firebase.auth
-                    .signInWithCredential(credential)
-                    .addOnCompleteListener { task ->
-                        if (task.isSuccessful) {
-                            Log.d(TAG, "Firebase signInWithGoogle: success")
-                        } else {
-                            Log.w(TAG, "Firebase signInWithGoogle: failure", task.exception)
-                        }
-                        continuation.resume(Unit)
-                    }
-            }
-        }
-
-        /**
-         * Refresh the AuthState based on the Firebase auth state.
-         *
-         * Get the ID token and other user information from Firebase
-         * and emit them with [_authState].
-         */
-        override suspend fun refreshFirebaseAuthState(): AuthState {
-            try {
-                val currentUser = Firebase.auth.currentUser ?: return AuthState.Unauthenticated.commit()
-                val idToken: FirebaseIdToken? =
-                    suspendCancellableCoroutine { continuation ->
-                        currentUser
-                            .getIdToken(true)
-                            .addOnSuccessListener { result ->
-                                Log.d(TAG, "Firebase ID Token: ${result.token}")
-                                continuation.resume(
-                                    result.token?.let {
-                                        FirebaseIdToken(
-                                            idToken = it,
-                                            exp = result.expirationTimestamp,
-                                        )
-                                    },
-                                )
-                            }.addOnFailureListener { exception ->
-                                Log.e(TAG, "Failed to get Firebase ID Token", exception)
-                                continuation.resume(null)
-                            }
-                    }
-                if (idToken == null) {
-                    return AuthState.Unauthenticated.commit()
-                }
-                return AuthState
-                    .Authenticated(
-                        user =
-                            User(
-                                name = DisplayName(currentUser.displayName ?: ""),
-                                email = Email(currentUser.email ?: ""),
-                                idToken = idToken,
-                            ),
-                    ).commit()
-            } catch (e: Exception) {
-                return AuthState.Unauthenticated.commit()
-            }
-        }
-
-        /**
-         * Commit the new AuthState and handle any side effects.
-         */
-        private suspend fun AuthState.commit(): AuthState {
-            Log.d(TAG, "AuthState.commit(): $this")
-            when (this) {
-                is AuthState.Authenticated ->
-                    appLoggerRepository.log(AppLoggerKeys.USER_AUTHENTICATED)
-
-                AuthState.Unauthenticated ->
-                    appLoggerRepository.log(AppLoggerKeys.USER_UNAUTHENTICATED)
-
-                AuthState.Unknown ->
-                    appLoggerRepository.log(AppLoggerKeys.USER_AUTH_UNKNOWN)
-            }
-            return this.also {
-                _authState.value = it
-            }
-        }
-
-        override suspend fun signOut() {
-            AuthState.Unauthenticated.commit()
-            try {
-                Firebase.auth.signOut()
-            } catch (e: Exception) {
-                Log.e(TAG, e.toString())
-            }
+    init {
+        CoroutineScope(Dispatchers.IO).launch {
+            refreshFirebaseAuthState()
         }
     }
 
-@Module
-@InstallIn(SingletonComponent::class)
-@Suppress("unused")
-abstract class AuthRepositoryModule {
-    @Binds
-    abstract fun bindAuthRepository(authRepository: AuthRepositoryImpl): AuthRepository
+    /**
+     * Sign in with Google ID Token and update the [authState].
+     */
+    override suspend fun signInWithGoogle(idToken: GoogleIdToken): AuthState {
+        firebaseSignInWithGoogle(idToken)
+        return refreshFirebaseAuthState()
+    }
+
+    /**
+     * Use the Google ID Token to sign in with Firebase.
+     */
+    private suspend fun firebaseSignInWithGoogle(idToken: GoogleIdToken) {
+        suspendCoroutine { continuation ->
+            Log.d(TAG, "firebaseAuthWithGoogle")
+            val credential = GoogleAuthProvider.getCredential(idToken.asString(), null)
+            Firebase.auth
+                .signInWithCredential(credential)
+                .addOnCompleteListener { task ->
+                    if (task.isSuccessful) {
+                        Log.d(TAG, "Firebase signInWithGoogle: success")
+                    } else {
+                        Log.w(TAG, "Firebase signInWithGoogle: failure", task.exception)
+                    }
+                    continuation.resume(Unit)
+                }
+        }
+    }
+
+    /**
+     * Refresh the AuthState based on the Firebase auth state.
+     *
+     * Get the ID token and other user information from Firebase
+     * and emit them with [_authState].
+     */
+    override suspend fun refreshFirebaseAuthState(): AuthState {
+        try {
+            val currentUser = Firebase.auth.currentUser ?: return AuthState.Unauthenticated.commit()
+            val idToken: FirebaseIdToken? =
+                suspendCancellableCoroutine { continuation ->
+                    currentUser
+                        .getIdToken(true)
+                        .addOnSuccessListener { result ->
+                            Log.d(TAG, "Firebase ID Token: ${result.token}")
+                            continuation.resume(
+                                result.token?.let {
+                                    FirebaseIdToken(
+                                        idToken = it,
+                                        exp = result.expirationTimestamp,
+                                    )
+                                },
+                            )
+                        }.addOnFailureListener { exception ->
+                            Log.e(TAG, "Failed to get Firebase ID Token", exception)
+                            continuation.resume(null)
+                        }
+                }
+            if (idToken == null) {
+                return AuthState.Unauthenticated.commit()
+            }
+            return AuthState
+                .Authenticated(
+                    user =
+                        User(
+                            name = DisplayName(currentUser.displayName ?: ""),
+                            email = Email(currentUser.email ?: ""),
+                            idToken = idToken,
+                        ),
+                ).commit()
+        } catch (e: Exception) {
+            return AuthState.Unauthenticated.commit()
+        }
+    }
+
+    /**
+     * Commit the new AuthState and handle any side effects.
+     */
+    private suspend fun AuthState.commit(): AuthState {
+        Log.d(TAG, "AuthState.commit(): $this")
+        when (this) {
+            is AuthState.Authenticated ->
+                appLoggerRepository.log(AppLoggerKeys.USER_AUTHENTICATED)
+
+            AuthState.Unauthenticated ->
+                appLoggerRepository.log(AppLoggerKeys.USER_UNAUTHENTICATED)
+
+            AuthState.Unknown ->
+                appLoggerRepository.log(AppLoggerKeys.USER_AUTH_UNKNOWN)
+        }
+        return this.also {
+            _authState.value = it
+        }
+    }
+
+    override suspend fun signOut() {
+        AuthState.Unauthenticated.commit()
+        try {
+            Firebase.auth.signOut()
+        } catch (e: Exception) {
+            Log.e(TAG, e.toString())
+        }
+    }
 }
 
 private const val TAG = "AuthRepository"
