+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/coroutines/DispatcherProvider.kt
@@ -17,14 +17,8 @@
 
 package com.chriscartland.garage.coroutines
 
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
-import javax.inject.Inject
-import javax.inject.Singleton
 
 interface DispatcherProvider {
     val main: CoroutineDispatcher
@@ -32,19 +26,8 @@ interface DispatcherProvider {
     val default: CoroutineDispatcher
 }
 
-@Singleton
-class DefaultDispatcherProvider
-    @Inject
-    constructor() : DispatcherProvider {
-        override val main: CoroutineDispatcher = Dispatchers.Main
-        override val io: CoroutineDispatcher = Dispatchers.IO
-        override val default: CoroutineDispatcher = Dispatchers.Default
-    }
-
-@Module
-@InstallIn(SingletonComponent::class)
-@Suppress("unused")
-abstract class DispatcherModule {
-    @Binds
-    abstract fun bindDispatcherProvider(impl: DefaultDispatcherProvider): DispatcherProvider
+class DefaultDispatcherProvider : DispatcherProvider {
+    override val main: CoroutineDispatcher = Dispatchers.Main
+    override val io: CoroutineDispatcher = Dispatchers.IO
+    override val default: CoroutineDispatcher = Dispatchers.Default
 }
