+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/internet/RetrofitNetworkConfigDataSource.kt
@@ -20,61 +20,45 @@ package com.chriscartland.garage.internet
 import android.util.Log
 import com.chriscartland.garage.data.NetworkConfigDataSource
 import com.chriscartland.garage.domain.model.ServerConfig
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Inject
-import javax.inject.Singleton
 
-class RetrofitNetworkConfigDataSource
-    @Inject
-    constructor(
-        private val network: GarageNetworkService,
-    ) : NetworkConfigDataSource {
-        override suspend fun fetchServerConfig(serverConfigKey: String): ServerConfig? {
-            try {
-                val response = network.getServerConfig(serverConfigKey)
-                if (response.code() != 200) {
-                    Log.e(TAG, "Response code is ${response.code()}")
-                    return null
-                }
-                val body = response.body()?.body
-                if (body == null) {
-                    Log.e(TAG, "Response body is null")
-                    return null
-                }
-                if (body.buildTimestamp.isNullOrEmpty()) {
-                    Log.e(TAG, "buildTimestamp is empty")
-                    return null
-                }
-                val remoteButtonBuildTimestamp = body.remoteButtonBuildTimestamp
-                if (remoteButtonBuildTimestamp.isNullOrEmpty()) {
-                    Log.e(TAG, "remoteButtonBuildTimestamp is empty")
-                    return null
-                }
-                if (body.remoteButtonPushKey.isNullOrEmpty()) {
-                    Log.e(TAG, "remoteButtonPushKey is empty")
-                    return null
-                }
-                return ServerConfig(
-                    buildTimestamp = body.buildTimestamp,
-                    remoteButtonBuildTimestamp = remoteButtonBuildTimestamp,
-                    remoteButtonPushKey = body.remoteButtonPushKey,
-                )
-            } catch (e: Exception) {
-                Log.e(TAG, "Error fetching server config: $e")
+class RetrofitNetworkConfigDataSource(
+    private val network: GarageNetworkService,
+) : NetworkConfigDataSource {
+    override suspend fun fetchServerConfig(serverConfigKey: String): ServerConfig? {
+        try {
+            val response = network.getServerConfig(serverConfigKey)
+            if (response.code() != 200) {
+                Log.e(TAG, "Response code is ${response.code()}")
                 return null
             }
+            val body = response.body()?.body
+            if (body == null) {
+                Log.e(TAG, "Response body is null")
+                return null
+            }
+            if (body.buildTimestamp.isNullOrEmpty()) {
+                Log.e(TAG, "buildTimestamp is empty")
+                return null
+            }
+            val remoteButtonBuildTimestamp = body.remoteButtonBuildTimestamp
+            if (remoteButtonBuildTimestamp.isNullOrEmpty()) {
+                Log.e(TAG, "remoteButtonBuildTimestamp is empty")
+                return null
+            }
+            if (body.remoteButtonPushKey.isNullOrEmpty()) {
+                Log.e(TAG, "remoteButtonPushKey is empty")
+                return null
+            }
+            return ServerConfig(
+                buildTimestamp = body.buildTimestamp,
+                remoteButtonBuildTimestamp = remoteButtonBuildTimestamp,
+                remoteButtonPushKey = body.remoteButtonPushKey,
+            )
+        } catch (e: Exception) {
+            Log.e(TAG, "Error fetching server config: $e")
+            return null
         }
     }
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkConfigDataSourceModule {
-    @Provides
-    @Singleton
-    fun provideNetworkConfigDataSource(network: GarageNetworkService): NetworkConfigDataSource = RetrofitNetworkConfigDataSource(network)
 }
 
 private const val TAG = "RetrofitNetworkConfig"
